mysite/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-positions==0.6.0
djangorestframework==3.12.2
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
